Update the panel location when its height changes
When the font scale changes during a session, so does the panel height.
This commit makes the panel position update properly when its height
changes for any reason.

Relatedly but differently, if the font scale is greater than 1.0 at login
under gnome-shell 3.20, there seems to be a bug in gnome-shell whereby
PanelBox.height is double-corrected for the type size. This results in a
panel floating a few pixels above the bottom of the monitor area. There
seems to be a later notification about a height change which actually does
make sense, so this commit works around the login glitch too ☺